README.md
Personal Diary Management System is based on a concept to generate Personal Daily Records and to add records & update them. Here User can add their daily details safely and it’s not time-consuming. Each Section contains high security so there is no chance of data loss. There’s maximum privacy to the recorded files. This System makes it easy to store daily records. The whole project is designed in ‘C’ language and different variables and strings have been used for the development of this project. This mini project is easy to operate and understand by the users. This is a simple mini project in C, built as a console application without using any graphics features. It’s easy to operate and understand by users. There is no error and warning content in the project. The design is so simple that users won’t find it difficult to use and navigate